// src/types/api/Activity.ts
/* eslint-disable */
/* @ts-ignore */
/**
 * DO NOT MODIFY IT BY HAND.
 * This file was automatically generated.
 */

import { Method } from './enums';

export interface Activity {
  amount?: string;
  currency?: string;
  date?: string;
  fromAddress?: string;
  fromId?: string;
  fromImage?: string;
  fromName?: string;
  id?: number;
  isViewed?: boolean;
  method?: Method;
  price?: string;
  toAddress?: string;
  toId?: string;
  toImage?: string;
  toName?: string;
  tokenId?: string;
  tokenImage?: string;
  tokenName?: string;
  txHash?: string;
}